+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * feat(ProseAccordionItem): new component
 * feat(ProseBadge): new component
 * feat(ProseKbd): new component
+* feat(ProseSteps): new component
 * feat(Banner): new component
 * feat(Card): new component
 * feat(CheckboxGroup): new component